refactor: SubwayLineBadge class 분리
